解析基金数据

+++ src/main/java/com/weizu/pojo/fund/FundNetWorthBean.java
@@ -14,7 +14,7 @@ public class FundNetWorthBean implements Serializable {
     /** 基金id */
     private long fundId;
     /** 单位净值 */
-    private Float netWorth;
+    private BigDecimal netWorth;
     /** 日期-天 */
     private String worthDate;
     /** 日涨幅 */
@@ -30,8 +30,8 @@ public class FundNetWorthBean implements Serializable {
     public void setId(Long id) {this.id = id;}
     public long getFundId() {return fundId;}
     public void setFundId(Long fundId) {this.fundId = fundId;}
-    public Float getNetWorth() {return netWorth;}
-    public void setNetWorth(Float netWorth) {this.netWorth = netWorth;}
+    public BigDecimal getNetWorth() {return netWorth;}
+    public void setNetWorth(BigDecimal netWorth) {this.netWorth = netWorth;}
     public String getWorthDate() {return worthDate;}
     public void setWorthDate(String worthDate) {this.worthDate = worthDate;}
     public BigDecimal getDayGrowth() {return dayGrowth;}
